Remove obsolete `DatasetManager` class
Previous functionality of this class is now primarily handled by `LocalDatastoreFile`,
global `TableConfig` objects, and `ETLManager`.
